Make AvxFftPlanner: Send+Sync (#55)
* Use Arc instead of Rc for Recipe in FftPlannerScalar

This is to make FftPlannerScalar: Send

* Make AvxFftPlanner: Send

This is done by making AvxPlannerInternalAPI: Send, which adds no
effective additional constraint since the only implementer of this trait
is AvxPlannerInternal, which is Send.

A test is added to ensure that FftPlanner: Send.

Why making FftPlanner: Send?
----------------------------------

Making FftPlanner: Send simplifies the usage of a single FftPlanner
in a large chunk of an multi-threaded application, hence benefitting of
the cache across threads.
This allows for instance to use the FftPlanner in a Mutex, or in a
global variable (e.g. using lazy_static!), which is may be useful even
if the application is actually not multi-threaded.

There is currently no performance cost in making FftPlanner: Send,
and other parts of the APIs already pay some cost in order to allow
multi-threading (e.g. FftPlanner.plan_ftt that returns an Arc).

Co-authored-by: Gaëtan Cassiers <gaetan.cassiers@uclouvain.be>