Remove unnecessary num_traits re-export
Nothing in the external API requires the consumer to use num_traits
directly. Removing this will decouple the rustfft major version from
the the num_traits major version.

FftNum is defined in terms of the `FromPrimitive` and `Signed` traits in the
num_traits crate however, so whenever they undergo a breaking change,
rustfft will still require a major version change.

Since this is a possible breaking change for consumers that happen to
consume num_traits via rustfft, this would require a major version
bump.